pyproject.toml
[tool.black]
include = '\.pyi?$'
line-length = 100

[tool.isort]
ensure_newline_before_comments = true
include_trailing_comma = true
line_length = 100
multi_line_output = 3
use_parentheses = true

[tool.pylint]
max-line-length = 100

[tool.pyright]
enableTypeIgnoreComments = true
reportMissingTypeStubs = false
typeCheckingMode = "basic"

[tool.pylic]
ignore_packages = [
  # only install these as part of dev tools to avoid GPL usage in main application
  "astroid",
  "pylint",
]
safe_licenses = [
  "3-Clause BSD License",
  "Apache License 2.0",
  "Apache Software License",
  "Artistic License",
  "BSD License",
  "BSD-derived (http://www.repoze.org/LICENSE.txt)",
  "BSD",
  "GNU Library or Lesser General Public License (LGPL)",
  "MIT License",
  "MIT",
  "Mozilla Public License 2.0 (MPL 2.0)",
  "Public Domain",
  "Python Software Foundation License",
]

[tool.pytest.ini_options]
addopts = """
    -W ignore::DeprecationWarning --no-header --cov --cov-report=html -vv
"""

[tool.poetry]
authors = [
  'Amy Troschinetz <spellbot@lexicalunit.com>',
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: End Users/Desktop",
  "Topic :: Communications :: Chat",
  "Topic :: Games/Entertainment :: Board Games",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.9",
]
description = 'The Discord bot for SpellTable'
homepage = 'http://spellbot.io/'
keywords = ['discord', 'magic', 'bot', 'mtg', 'SpellTable']
license = 'MIT'
name = 'spellbot'
packages = [
  {include = "spellbot", from = "src"},
]
readme = 'README.md'
repository = 'https://github.com/lexicalunit/spellbot'
version = "7.12.5"

[tool.poetry.dependencies]
# discord.py depends on "aiohttp<3.8.0" yet works just fine with later versions.
# poetry won't let us install the latest aiohttp due to this dependency, so in
# order to install newer versions of aiohttp we have to drop down to pip:
# $ pip install -U "aiohttp==3.8.1"
# $ pip install -U "pytest-aiohttp==1.0.4"
# So we do this in .github/workflows/ci.yaml, Dockerfile, and scripts/setup.sh.
Babel = "^2.10.3"
PyYAML = "^6.0"
SQLAlchemy = "^1.4.39"
SQLAlchemy-Utils = "^0.38.2"
aiohttp-jinja2 = "^1.5"
aiohttp-retry = "^2.4.8"
alembic = "^1.8.0"
asgiref = "^3.5.2"
click = "^8.1.3"
coloredlogs = "^15.0.1"
datadog = "^0.44.0"
ddtrace = "^1.2.1"
discord-py-interactions = "^3.0.2"
"discord.py" = "1.7.3"
dunamai = "^1.12.0"
expiringdict = "^1.2.2"
gunicorn = "^20.1.0"
humanize = "^4.2.1"
hupper = "^1.10.3"
importlib-resources = "^5.8.0"
psycopg2-binary = "^2.9.3"
pygicord = "^1.0.1"
python = '>=3.9,<4'
python-dateutil = "^2.8.2"
python-dotenv = "^0.20.0"
pytz = "^2022.1"
supervisor = "^4.2.4"
types-PyYAML = "^6.0.4"
types-python-dateutil = "^2.8.9"
types-pytz = "^2022.1.1"
types-toml = "^0.10.4"
uvloop = "^0.16.0"
wrapt = "^1.14.1"

[tool.poetry.dev-dependencies]
Faker = "^13.14.0"
GitPython = "^3.1.27"
black = "^22.3.0"
callee = "^0.3.1"
coverage = "^6.4.1"
factory-boy = "^3.2.1"
flake8 = "^4.0.1"
flake8-commas = "^2.1.0"
flake8-print = "^5.0.0"
isort = "^5.10.1"
pylic = "^3.1.0"
pylint = "2.14.0"
pyright = "^1.1.255"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
pytest-cov = "^3.0.0"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.8.1"
sqlalchemy-stubs = "^0.4"
syrupy = "^2.3.0"

[tool.poetry.scripts]
spellbot = 'spellbot:main'

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.4"]
